refactor blog repository
